Moved files around to fix previous test errors
