+++ tests/app/celery/test_process_ses_receipts_tasks.py
@@ -18,7 +18,7 @@ from app.notifications.notifications_ses_callback import (
     remove_emails_from_bounce,
     remove_emails_from_complaint,
 )
-from tests.app.aws import ses_complaint_callback
+from tests.app.aws.aws_mocks import ses_complaint_callback
 from tests.app.conftest import sample_notification as create_sample_notification
 from tests.app.db import create_notification, create_service_callback_api
